fix(TUP-18073): enhance and unify the component folders  (#1111)
* fix(TUP-18073): miss component zip in installed folder

* fix(TUP-18073): install lib in m2 always and unify the component folder

+++ main/plugins/org.talend.updates.runtime/src/main/java/org/talend/updates/runtime/nexus/component/ComponentsDeploymentManager.java
@@ -45,10 +45,12 @@ public class ComponentsDeploymentManager {
 
     private void moveToSharedFolder(File componentZipFile) throws IOException {
         File sharedCompFile = new File(PathUtils.getComponentsSharedFolder(), componentZipFile.getName());
-        FilesUtils.copyFile(componentZipFile, sharedCompFile);
-        boolean deleted = componentZipFile.delete();
-        if (!deleted) {// failed to delete in time
-            componentZipFile.deleteOnExit(); // try to delete when exit
+        if (!componentZipFile.equals(sharedCompFile)) { // not in same folder
+            FilesUtils.copyFile(componentZipFile, sharedCompFile);
+            boolean deleted = componentZipFile.delete();
+            if (!deleted) {// failed to delete in time
+                componentZipFile.deleteOnExit(); // try to delete when exit
+            }
         }
     }
